# public/api-docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Payment Portal API",
    "description": "API for managing payments, cards, and user accounts in the Payment Portal system",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@paymentportal.com"
    }
  },
  "servers": [
    {
      "url": "https://yourdomain.com/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8000/api",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/cards": {
      "get": {
        "tags": ["Cards"],
        "summary": "List user's cards",
        "description": "Get list of user's payment cards",
        "responses": {
          "200": {
            "description": "List of cards",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Cards"],
        "summary": "Create a new card",
        "description": "Add a new payment card",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Card created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/cards/{id}": {
      "get": {
        "tags": ["Cards"],
        "summary": "Get card details",
        "description": "Get specific card details",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Card ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Card details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "404": {
            "description": "Card not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Cards"],
        "summary": "Update card",
        "description": "Update card information",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Card ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Card updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "404": {
            "description": "Card not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Cards"],
        "summary": "Delete card",
        "description": "Delete a payment card",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Card ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Card deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "404": {
            "description": "Card not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/cards/search": {
      "get": {
        "tags": ["Cards"],
        "summary": "Search cards",
        "description": "Search user's cards",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Search query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/cards/{card}/set-default": {
      "post": {
        "tags": ["Cards"],
        "summary": "Set card as default",
        "description": "Set a card as the default payment method",
        "parameters": [
          {
            "name": "card",
            "in": "path",
            "required": true,
            "description": "Card ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Card set as default successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "404": {
            "description": "Card not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Card": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "card_number": {
            "type": "string"
          },
          "expiry_month": {
            "type": "integer"
          },
          "expiry_year": {
            "type": "integer"
          },
          "cvv": {
            "type": "string"
          },
          "cardholder_name": {
            "type": "string"
          },
          "is_default": {
            "type": "boolean"
          },
          "is_active": {
            "type": "boolean"
          },
          "user_id": {
            "type": "integer"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CardRequest": {
        "type": "object",
        "required": ["card_number", "expiry_month", "expiry_year", "cardholder_name"],
        "properties": {
          "card_number": {
            "type": "string",
            "minLength": 13,
            "maxLength": 19
          },
          "expiry_month": {
            "type": "integer",
            "minimum": 1,
            "maximum": 12
          },
          "expiry_year": {
            "type": "integer",
            "minimum": 2024
          },
          "cvv": {
            "type": "string"
          },
          "cardholder_name": {
            "type": "string",
            "maxLength": 255
          },
          "is_default": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "CardUpdateRequest": {
        "type": "object",
        "properties": {
          "expiry_month": {
            "type": "integer",
            "minimum": 1,
            "maximum": 12
          },
          "expiry_year": {
            "type": "integer",
            "minimum": 2024
          },
          "cvv": {
            "type": "string"
          },
          "cardholder_name": {
            "type": "string",
            "maxLength": 255
          },
          "is_default": {
            "type": "boolean"
          }
        }
      },
      "CardResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Card"
          }
        }
      },
      "CardsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Card"
            }
          }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  },
...
}